+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+### Version 0.1.1
+- Fix issue where double letters could produce false mismatches
+- Add two more 5 letter words
+
 ### Version 0.1.0
 - Initial release